Add a concept of ambigous KOD table
If the cracking heuristics can not completely deduct a KOD table, allow for unresolved entries with the value of -1

To fix up some of these entries, the user can provide a new fix for the table via the -f option.

To further guide the user, print out preliminarily decoded struc records and some stats about duplicated and missing KOD mappings.